+++ lib/src/_extensions/_index.dart
@@ -1,7 +1,0 @@
-// Copyright ©2022, GM Consult (Pty) Ltd
-// BSD 3-Clause License
-// All rights reserved
-
-// List all exports below in alphabetical order.
-export 'postings_extension.dart';
-export 'term_dictionary_extensions.dart';