+++ lib/src/_models/document.dart
@@ -9,7 +9,7 @@ import 'package:text_indexing/text_indexing.dart';
 ///
 /// The [Document] properties are stored in the [Postings] of the index,
 /// distributed over the vocabulary terms.  [Document]s are extracted from
-/// the [Postings] by filtering the [Postings] entries for [PostingsMap]
+/// the [Postings] by filtering the [Postings] entries for [DocumentPostingsEntry]
 /// elements with the same identifier [docId] as the document.
 abstract class Document {
   //
@@ -17,19 +17,18 @@ abstract class Document {
   /// Instantiates a const [Document] instance.
   const Document();
 
-  /// The document's unique identifier in the corpus.
-  String get docId;
+  /// The document's unique identifier ([DocId]) in the corpus.
+  DocId get docId;
 
-  /// An alphabetically ordered list of the terms in the document as they
-  /// appear in the index.
-  List<String> get terms;
+  /// An alphabetically ordered list of the [Term]s in the document.
+  List<Term> get terms;
 
-  /// A hash of terms to the number of times each term occurs in the document.
-  Map<String, int> get termFrequencies;
+  /// A hashmap of [Term]s to the number of times ([Ft]) each [Term] occurs in
+  /// the document.
+  Map<Term, Ft> get termFrequencies;
 
-  /// A hash of terms to a hash of the fields of the document in which they
-  /// appear and a list of term positions in that field.
-  Map<String, Map<String, List<int>>> get termFieldPostings;
+  /// A hashmap of [Term]s to [FieldPostings] for the document.
+  Map<FieldName, FieldPostings> get termFieldPostings;
 
   //
 }