+++ test/text_indexing_test.dart
@@ -43,9 +43,14 @@ void main() {
 
       indexer.postingsStream.listen((event) {
         if (event.isNotEmpty) {
-          final docId = event.first.docId;
-          final terms = event.map((e) => e.term).toList();
-          print('$docId: $terms');
+          final PostingsEntry posting = event.entries.first;
+          if (posting.value.isNotEmpty) {
+            final DocumentPostingsEntry docPostings =
+                posting.value.entries.first;
+            final docId = docPostings.docId;
+            final terms = event.terms;
+            print('$docId: $terms');
+          }
         }
       });
 
@@ -99,9 +104,14 @@ void main() {
 
       indexer.postingsStream.listen((event) {
         if (event.isNotEmpty) {
-          final docId = event.first.docId;
-          final terms = event.map((e) => e.term).toList();
-          print('$docId: $terms');
+          final PostingsEntry posting = event.entries.first;
+          if (posting.value.isNotEmpty) {
+            final DocumentPostingsEntry docPostings =
+                posting.value.entries.first;
+            final docId = docPostings.docId;
+            final terms = event.terms;
+            print('$docId: $terms');
+          }
         }
       });